+++ src/steps/account/index.ts
@@ -10,6 +10,7 @@ import { createAccountEntity } from './converters';
 export async function fetchAccount({
   instance,
   jobState,
+  logger,
 }: IntegrationStepExecutionContext<IntegrationConfig>) {
   const apiClient = createAPIClient(instance.config);
   const account = await apiClient.getAccount();
@@ -18,6 +19,12 @@ export async function fetchAccount({
       createAccountEntity(account),
     );
     await jobState.setData(ACCOUNT_ENTITY_DATA_KEY, accountEntity);
+  } else {
+    logger.publishEvent({
+      name: 'missing_account_entity',
+      description:
+        'Could not find account entity: please ensure that entered email address is correct.',
+    });
   }
 }