+++ CHANGELOG.md
@@ -8,10 +8,16 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Fixed
+
+- Add `missing account` handling
+
 ## [1.0.1] - 2022-01-28
 
 ### Fixed
-- Use `await` for `iterateApi` callback function to properly create all the user roles relationships
+
+- Use `await` for `iterateApi` callback function to properly create all the user
+  roles relationships
 
 ## [1.0.0] - 2022-01-14